+++ src/index.ts
@@ -1,13 +1,6 @@
 export { formatWxml, type FormatWxmlOptions } from './format'
 
-export {
-  defaultExport as default,
-  defaultExport,
-  languages,
-  options,
-  parsers,
-  printers,
-} from './plugin'
+export { defaultExport as default, defaultExport, languages, options, parsers, printers } from './plugin'
 
 export type { WxmlPluginOptions } from './plugin-options'